nv_client/requirements.txt
pycrypto
grpcio
grpcio_tools
wmi
pywin32
rsa
pyyaml
pyinstaller
psutil
fastapi
uvicorn
Cython
pillow